osd/ReplicatedBackend: cast variable to unsigned before comparing
after the yaml-to-cxx migration, osd_deep_scrub_stride is an OPT_SIZE,
and is hence represented with an uint64. so we need to cast `r` to
uint64_t before comparing it with this setting for silencing GCC
warning.

+++ src/osd/ReplicatedBackend.cc
@@ -671,7 +671,7 @@ int ReplicatedBackend::be_deep_scrub(
       pos.data_hash << bl;
     }
     pos.data_pos += r;
-    if (r == cct->_conf->osd_deep_scrub_stride) {
+    if (static_cast<uint64_t>(r) == cct->_conf->osd_deep_scrub_stride) {
       dout(20) << __func__ << "  " << poid << " more data, digest so far 0x"
 	       << std::hex << pos.data_hash.digest() << std::dec << dendl;
       return -EINPROGRESS;
